dashboard/requirements_dashboard.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
